added NewsPageClient page
